Adding app menu
